+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Version 0.2.1 2016-11-28
+
+- Use lyp finalizer API to display assertion summary on exit.
+
+## Version 0.2.0 2016-06-02
+
+- Fix assert-compare-fn messages.
+- Use new \require form with relative path.
+- Update with new lyp scheme interface.
+
 ## Version 0.1.3 2016-01-27
 
 - Improved assertion count.